Fixed bug of view popping from stack when title empty. Change some types to avoid casting errors. We should always use NSInteger instead of int since NSInteger is architecture independent (32 bit vs 64 bit on newer iPhones, etc).

+++ PQPTasks/AJVArithTestModel.h
@@ -12,7 +12,7 @@
 
 @property (nonatomic) NSInteger num;
 
-- (int)add:(NSInteger)addNum;
-- (int)mult:(NSInteger)multNum;
+- (NSInteger)add:(NSInteger)addNum;
+- (NSInteger)mult:(NSInteger)multNum;
 
 @end
